Compress XACT implementation to one file
